Move blas_connector to module_external
