+++ source/source_base/kernels/math_kernel_op.h
@@ -5,7 +5,7 @@
 
 #include "source_base/macros.h"
 
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/parallel_reduce.h"
 
 #include "source_base/module_device/memory_op.h"
