+++ source/source_base/kernels/test/math_kernel_test.cpp
@@ -1,4 +1,4 @@
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/constants.h"
 #include "source_base/module_device/memory_op.h"
 #include "source_base/kernels/math_kernel_op.h"
