+++ source/source_basis/module_ao/ORB_atomic_lm.cpp
@@ -1,7 +1,7 @@
 #include "ORB_atomic_lm.h"
 #include "source_base/sph_bessel_recursive.h"
 #include "source_io/module_parameter/parameter.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_base/math_integral.h"
 #include "source_base/math_sphbes.h"
