+++ source/source_basis/module_ao/test/CMakeLists.txt
@@ -7,9 +7,9 @@ list(APPEND depend_files
   ../../../source_base/math_ylmreal.cpp
   ../../../source_base/ylm.cpp
   ../../../source_base/memory.cpp
-  ../../../source_base/blas_connector_base.cpp
-  ../../../source_base/blas_connector_vector.cpp
-  ../../../source_base/blas_connector_matrix.cpp
+  ../../../source_base/module_external/blas_connector_base.cpp
+  ../../../source_base/module_external/blas_connector_vector.cpp
+  ../../../source_base/module_external/blas_connector_matrix.cpp
   ../../../source_base/complexarray.cpp
   ../../../source_base/complexmatrix.cpp
   ../../../source_base/matrix.cpp
