+++ source/source_basis/module_pw/test/CMakeLists.txt
@@ -4,7 +4,7 @@ AddTest(
   LIBS parameter ${math_libs} planewave device 
   SOURCES ../../../source_base/matrix.cpp ../../../source_base/complexmatrix.cpp ../../../source_base/matrix3.cpp ../../../source_base/tool_quit.cpp
           ../../../source_base/mymath.cpp ../../../source_base/timer.cpp ../../../source_base/memory.cpp
-          ../../../source_base/blas_connector_base.cpp ../../../source_base/blas_connector_vector.cpp ../../../source_base/blas_connector_matrix.cpp   
+          ../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp   
           ../../../source_base/libm/branred.cpp ../../../source_base/libm/sincos.cpp 
           ../../../source_base/module_device/memory_op.cpp
           depend_mock.cpp pw_test.cpp test1-1-1.cpp test1-1-2.cpp test1-2.cpp test1-3.cpp test1-4.cpp  test1-5.cpp
