+++ source/source_esolver/esolver_ks_lcao_tddft.cpp
@@ -10,7 +10,7 @@
 #include "source_io/output_log.h"
 
 //--------------temporary----------------------------
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/global_function.h"
 #include "source_base/lapack_connector.h"
 #include "source_base/scalapack_connector.h"
