+++ source/source_estate/math_tools.h
@@ -1,4 +1,4 @@
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/complexmatrix.h"
 #include "source_base/matrix.h"
 #include "source_base/scalapack_connector.h"
