+++ source/source_hamilt/module_xc/test/CMakeLists.txt
@@ -40,7 +40,7 @@ AddTest(
     ../../../source_base/memory.cpp
     ../../../source_base/libm/branred.cpp
     ../../../source_base/libm/sincos.cpp
-    ../../../source_base/blas_connector_base.cpp ../../../source_base/blas_connector_vector.cpp ../../../source_base/blas_connector_matrix.cpp
+    ../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp
     ../../../source_basis/module_pw/module_fft/fft_bundle.cpp
     ../../../source_basis/module_pw/module_fft/fft_cpu.cpp
     ${FFT_SRC}
@@ -73,7 +73,7 @@ AddTest(
     ../xc_functional_vxc.cpp
     ../xc_functional_libxc_vxc.cpp
     ../xc_functional_libxc_tools.cpp
-    ../../../source_base/blas_connector_base.cpp ../../../source_base/blas_connector_vector.cpp ../../../source_base/blas_connector_matrix.cpp
+    ../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp
     ../../../source_base/matrix.cpp
     ../../../source_base/memory.cpp
     ../../../source_base/timer.cpp
