+++ source/source_io/fR_overlap.cpp
@@ -4,7 +4,7 @@
 #include "source_base/timer.h"
 #include "source_base/tool_title.h"
 #include "source_base/math_integral.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 
 template <typename T>
 FR_overlap<T>::FR_overlap()
