+++ source/source_lcao/module_deepks/deepks_descriptor.cpp
@@ -8,7 +8,7 @@
 #include "deepks_descriptor.h"
 
 #include "LCAO_deepks_io.h" // mohan add 2024-07-22
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/constants.h"
 #include "source_base/libm/libm.h"
 #include "source_base/parallel_reduce.h"
