+++ source/source_lcao/module_gint/gint_gamma_vl.cpp
@@ -4,7 +4,7 @@
 #include "gint_gamma.h"
 #include "gint_tools.h"
 #include "grid_technique.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/memory.h"
 #include "source_base/timer.h"
 #include "source_basis/module_ao/ORB_read.h"
