+++ source/source_lcao/module_gint/gint_vl_old.cpp
@@ -5,7 +5,7 @@
 #include "grid_technique.h"
 #include "source_base/ylm.h"
 #include "source_pw/module_pwdft/global.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_base/array_pool.h"
 #include "source_base/vector3.h"
