+++ source/source_lcao/module_gint/kernels/cuda/gemm_selector.cu
@@ -3,7 +3,7 @@
 #include "gemm_selector.cuh"
 #include "vbatch_matrix_mul.cuh"
 #include "cuda_tools.cuh"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "code_gen.cuh"
 
 /*
