+++ source/source_lcao/module_gint/mult_psi_dmr.cpp
@@ -1,7 +1,7 @@
 #include "gint_tools.h"
 #include "source_base/timer.h"
 #include "source_base/ylm.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 
 namespace Gint_Tools{
