+++ source/source_lcao/module_gint/temp_gint/gint_vl_metagga.cpp
@@ -1,4 +1,4 @@
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "gint_common.h"
 #include "gint_vl_metagga.h"
 #include "phi_operator.h"
