+++ source/source_lcao/module_gint/temp_gint/gint_vl_nspin4.cpp
@@ -1,6 +1,6 @@
 #include "source_base/global_function.h"
 #include "source_lcao/module_hcontainer/hcontainer.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "gint_common.h"
 #include "gint_vl_nspin4.h"
 #include "phi_operator.h"
