+++ source/source_lcao/module_gint/temp_gint/phi_operator.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "phi_operator.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/global_function.h"
 
 namespace ModuleGint
