+++ source/source_lcao/module_hcontainer/atom_pair.cpp
@@ -1,7 +1,7 @@
 #include "atom_pair.h"
 #include <complex>
 #include <cassert>
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 
 namespace hamilt
 {
