+++ source/source_lcao/module_lr/ao_to_mo_transformer/ao_to_mo_serial.cpp
@@ -1,5 +1,5 @@
 #include "ao_to_mo.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/tool_title.h"
 #include "source_lcao/module_lr/utils/lr_util.h"
 namespace LR
