+++ source/source_lcao/module_lr/dm_trans/dm_trans_serial.cpp
@@ -1,5 +1,5 @@
 #include "dm_trans.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/tool_title.h"
 #include "source_base/global_function.h"
 #include "source_lcao/module_lr/utils/lr_util.h"
