+++ source/source_lcao/module_lr/operator_casida/operator_lr_hxc.cpp
@@ -1,7 +1,7 @@
 #include "operator_lr_hxc.h"
 #include <vector>
 #include "source_io/module_parameter/parameter.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_lcao/module_lr/utils/lr_util.h"
 #include "source_lcao/module_lr/utils/lr_util_hcontainer.h"
