+++ source/source_lcao/module_operator_lcao/dspin_lcao.cpp
@@ -1,6 +1,6 @@
 #include "dspin_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_base/memory.h"
 #include "source_base/tool_title.h"
