+++ source/source_lcao/module_ri/module_exx_symmetry/symmetry_rotation_R_hcontainer.hpp
@@ -1,5 +1,5 @@
 #include "symmetry_rotation.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/parallel_reduce.h"
 namespace ModuleSymmetry
 {
