+++ source/source_pw/module_pwdft/hamilt_pw.cpp
@@ -1,7 +1,7 @@
 #include "hamilt_pw.h"
 
 #include "source_io/module_parameter/parameter.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/global_function.h"
 #include "source_base/global_variable.h"
 #include "source_pw/module_pwdft/global.h"
