+++ source/source_pw/module_pwdft/operator_pw/nonlocal_pw.cpp
@@ -1,7 +1,7 @@
 #include "nonlocal_pw.h"
 
 #include "source_io/module_parameter/parameter.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/tool_quit.h"
