+++ source/source_pw/module_pwdft/operator_pw/onsite_proj_pw.cpp
@@ -1,6 +1,6 @@
 #include "onsite_proj_pw.h"
 
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/timer.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/tool_quit.h"
