+++ source/source_pw/module_pwdft/operator_pw/op_exx_pw.h
@@ -2,7 +2,7 @@
 #define OPEXXPW_H
 
 #include "operator_pw.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/kernels/math_kernel_op.h"
 #include "source_base/macros.h"
 #include "source_base/matrix.h"
