+++ source/source_pw/module_pwdft/test/CMakeLists.txt
@@ -15,7 +15,7 @@ AddTest(
 	../../../source_base/global_file.cpp
 	../../../source_base/memory.cpp
 	../../../source_base/timer.cpp
-	../../../source_base/blas_connector_base.cpp ../../../source_base/blas_connector_vector.cpp ../../../source_base/blas_connector_matrix.cpp
+	../../../source_base/module_external/blas_connector_base.cpp ../../../source_base/module_external/blas_connector_vector.cpp ../../../source_base/module_external/blas_connector_matrix.cpp
 	../../../source_base/parallel_global.cpp
 	../../../source_base/parallel_comm.cpp
 	../../../source_base/parallel_common.cpp
