+++ source/source_pw/module_stodft/sto_che.cpp
@@ -1,5 +1,5 @@
 #include "sto_che.h"
-#include "source_base/blas_connector.h"
+#include "source_base/module_external/blas_connector.h"
 #include "source_base/module_device/device.h"
 #include "source_base/kernels/math_kernel_op.h"
 #include "source_base/module_container/ATen/kernels/blas.h"
